# docs/swagger.yaml
definitions:
  api.JsonObject:
    additionalProperties: true
    type: object
  dao.DeviceStatistics:
    properties:
      date:
        description: 日期
        type: string
      disk_usage:
        description: 磁盘使用率
        type: number
      downstream_traffic:
        description: 下行流量
        type: number
      income:
        description: 收益
        type: number
      latency:
        description: 时延
        type: number
      nat_ratio:
        description: NAT率
        type: number
      online_time:
        description: 在线时长
        type: number
      pkg_loss_ratio:
        description: 丢包率
        type: number
      retrieval_count:
        description: 检索数量
        type: number
      upstream_traffic:
        description: 上行流量
        type: number
    type: object
  dao.UserDeviceProfile:
    properties:
      abnormal_num:
        description: 设备异常
        type: integer
      cumulative_profit:
        description: 累计收益
        type: number
      month_profit:
        description: 近30天收益
        type: number
      offline_num:
        description: 设备离线
        type: integer
      online_num:
        description: 在线设备数
        type: integer
      seven_days_profit:
        description: 近7天收益
        type: number
      today_profit:
        description: 今日收益
        type: number
      total_bandwidth:
        description: 总上行速度
        type: number
      total_num:
        description: 设备总数
        type: integer
      yesterday_profit:
        description: 昨日收益
        type: number
    type: object
  errors.GenericError:
    properties:
      code:
        type: integer
      err: {}
    type: object
...
  model.DeviceInfo:
    properties:
      active_status:
        description: 激活状态
        type: integer
      bandwidth_down:
        description: 下行带宽
        type: number
      bandwidth_up:
        description: 上行带宽
        type: number
      bind_status:
        description: 绑定状态
        type: string
      block_count:
        description: blocks
        type: integer
      bound_at:
        description: 绑定时间
        type: string
      cpu_cores:
        description: CPU核数
        type: integer
      cpu_usage:
        description: CPU使用率
        type: number
      created_at:
        type: string
      cumulative_profit:
        description: 总收益
        type: number
      deleted_at:
        type: string
      device_id:
        description: 设备ID
        type: string
      device_name:
        description: 设备名称
        type: string
      device_rank:
        description: 排名
        type: integer
      device_status:
        description: 设备状态
        type: string
      disk_space:
        description: 磁盘大小
        type: number
      disk_type:
        description: 磁盘类型
        type: string
      disk_usage:
        description: 磁盘使用率
        type: number
      download_count:
        description: 检索次数
        type: integer
      external_ip:
        description: 外网IP
        type: string
      id:
        type: integer
      internal_ip:
        description: 内网IP
        type: string
      io_system:
        description: 文件系统
        type: string
      ip_city:
        description: 城市
        type: string
      ip_country:
        description: 国家
        type: string
      ip_location:
        description: 国家地区
        type: string
      ip_province:
        description: 省
        type: string
      latency:
        description: 线路时延
        type: number
      latitude:
        description: 纬度
        type: number
      longitude:
        description: 经度
        type: number
      mac_location:
        description: mac地址
        type: string
      memory:
        description: 内存
        type: number
      memory_usage:
        description: 内存使用率
        type: number
      month_profit:
        description: 月收益
        type: number
      nat_ratio:
        description: NAT率
        type: number
      nat_type:
        description: NAT 类型
        type: string
      network_info:
        description: 网络信息
        type: string
      network_type:
        description: 网络类型
        type: string
      node_type:
        description: 类型
        type: integer
      online_time:
        description: 在线时长
        type: number
      operator:
        description: 操作系统
        type: string
      pkg_loss_ratio:
        description: 丢包率
        type: number
      product_type:
        description: 产品类型
        type: string
      scheduler_id:
        description: schedulerID
        type: string
      seven_days_profit:
        description: 七天收益
        type: number
      sn_code:
        description: sn码
        type: string
      system_version:
        description: 版本
        type: string
      today_online_time:
        description: 今日在线时长
        type: number
      today_profit:
        description: 今日收益
        type: number
      total_download:
        description: 总下行流量
        type: number
      total_upload:
        description: 总上行流量
        type: number
      updated_at:
        type: string
      upnp:
        description: UPNP
        type: string
      user_id:
        description: 用户ID（钱包地址）
        type: string
      work_status:
        description: 设备诊断状态
        type: string
      yesterday_profit:
        description: 昨日收益
        type: number
    type: object
  model.FullNodeInfo:
    properties:
      candidate_count:
        description: L1 候选节点
        type: integer
      created_at:
        type: string
      edge_count:
        description: L2 边缘节点
        type: integer
      id:
        type: integer
      next_election_time:
        description: 下一轮选举时间
        type: string
      retrieval_count:
        description: 加速次数
        type: integer
      time:
        type: string
      total_carfile:
        description: 已加速Carflie
        type: integer
      total_carfile_size:
        description: 已加速Carflie大小
        type: number
      total_downstream_bandwidth:
        description: 下载带宽
        type: number
      total_node_count:
        description: 全球分布式节点数量
        type: integer
      total_storage:
        description: 存储容量
        type: number
      total_upstream_bandwidth:
        description: 上行带宽
        type: number
      updated_at:
        type: string
      validator_count:
        description: L1 验证节点
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /all_areas:
    get:
      consumes:
      - application/json
      description: GetAllAreas is a handler to get a slice of record(s)from device
        info table in database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取所有的地区列表
      tags:
      - 矿工页
...
  /get_device_info:
    get:
      consumes:
      - application/json
      description: GetDeviceInfoHandler is a handler to get a slice of record(s) from
        device info table in database
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 设备ID
        in: query
        name: device_id
        type: string
      - description: 国家地区
        in: query
        name: ip_location
        type: integer
      - description: 节点类型，默认全部
        in: query
        name: node_type
        type: integer
      - description: 分页页码，默认 1
        in: query
        name: page
        type: integer
      - description: 分页页数 默认 50
        in: query
        name: page_size
        type: integer
      - description: 排序， Enums(ASC, DESC)
        enum:
        - ASC
        - DESC
        in: query
        name: order
        type: string
      - description: 排序字段
        in: query
        name: order_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                list:
                  items:
                    $ref: '#/definitions/model.DeviceInfo'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取矿工页列表
      tags:
      - 矿工页
  /get_diagnosis_days:
    get:
      consumes:
      - application/json
      description: GetDeviceDiagnosisDailyHandler is a function to get a slice of
        record(s) from device info daily diagnosis
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 开始时间
        in: query
        name: from
        type: string
      - description: 结束时间
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                series_data:
                  $ref: '#/definitions/dao.DeviceStatistics'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取历史数据诊断
      tags:
      - 矿工页
  /get_diagnosis_hours:
    get:
      consumes:
      - application/json
      description: GetDeviceDiagnosisHourHandler is a function to get a slice of record(s)
        from device info hourly diagnosis
      parameters:
      - description: 设备ID
        in: query
        name: device_id
        required: true
        type: string
      - description: 开始时间
        in: query
        name: from
        type: string
      - description: 结束时间
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                series_data:
                  $ref: '#/definitions/dao.DeviceStatistics'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取实时监控
      tags:
      - 矿工页
  /get_index_info:
    get:
      consumes:
      - application/json
      description: GetIndexInfoHandler is a function to get a single record from the
        full node info table in the redis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.FullNodeInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 获取数据总览数据
      tags:
      - 首页
...
  /get_user_device_profile:
    get:
      consumes:
      - application/json
      description: DeviceBindingHandler is a function to get user devices overview
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.JsonObject'
            - properties:
                profile:
                  $ref: '#/definitions/dao.UserDeviceProfile'
                series_data:
                  items:
                    $ref: '#/definitions/dao.DeviceStatistics'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.GenericError'
      summary: 用户总览
      tags:
      - 控制台
...
swagger: "2.0"